requirements.txt
streamlit==1.13.0
numpy==1.19.5
pandas==1.4.3
Pillow==9.0.0